Changed score vaiables to use int instead of double. I thought it would make things easier until I can get Player::save() working correctly.

+++ game/Player.h
@@ -17,14 +17,14 @@ public:
     void heal(int change);
     void award(int pts);
     int getHealth();
-    double getScore();
-    double getAllTime();
+    int getScore();
+    int getAllTime();
 private:
     int id;
     std::string name;
     int health;
-    double score;
-    double allTimeScore;
+    int score;
+    int allTimeScore;
 };
 
 #endif /* PLAYER_H */
